feat: add batch execution capability to worker pool, allowing multiple tasks to be processed simultaneously and improving task management efficiency

+++ src/config/faber.rs
@@ -35,5 +35,5 @@ pub struct FaberConfigOverrides {
     pub host: Option<String>,
     pub port: Option<u16>,
     pub auth_enabled: Option<bool>,
-    pub workers: Option<usize>,
+    pub workers: Option<u16>,
 }
